+++ src/hooks/useArticle.ts
@@ -1,19 +1,6 @@
 import getArticles from "@/services/getArticles";
+import { Articles } from "@/types/article";
 import { useCallback, useEffect, useState } from "react";
-
-export interface Articles {
-  id: number;
-  title: string;
-  content: string;
-  image: string;
-  likeCount: number;
-  createdAt: string;
-  updatedAt: string;
-  writer: {
-    nickname: string;
-    id: number;
-  };
-}
 
 interface Props {
   setArticleResults: React.Dispatch<React.SetStateAction<Articles[] | null>>;
