+++ src/types/article.ts
@@ -1,0 +1,13 @@
+export interface Articles {
+  id: number;
+  title: string;
+  content: string;
+  image: string;
+  likeCount: number;
+  createdAt: string;
+  updatedAt: string;
+  writer: {
+    nickname: string;
+    id: number;
+  };
+}
